+++ CHANGELOG.md
@@ -7,6 +7,8 @@ All notable changes to this project will be documented in this file. See [conven
 
 ### Features
 
+- add repeatable custom template variables with `--var key=value`
+- modularize the CLI and add non-interactive generation flags
 - gogen project finished - ([95041ad](https://github.com/qiaopengjun5162/gogen/commit/95041adb4bb0dbae7520dc9e799c38342af0df21)) - QiaoPengjun5162
 
 ### Miscellaneous Chores